requirements.txt
aiohttp
aiosignal
async-timeout
attrs
blinker
cattrs
charset-normalizer
click
colorama
exceptiongroup
flask
frozenlist
gunicorn 
highrise-bot-sdk 
idna
itsdangerous 
jinja2 
markupsafe 
multidict 
pendulum
python-dateutil 
python-dotenv 
pytzdata 
quattro 
six
typing-extensions 
werkzeug 
yarl 